% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPfinder.R
\name{bpFinder}
\alias{bpFinder}
\title{Find and store any dates that are mentioned in random locations by looking for BP}
\usage{
bpFinder(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to save miscellaneous dates
}
\examples{
bpFinder()
}
\keyword{BP}
