% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanString.R
\name{cleanString}
\alias{cleanString}
\title{Text processing to allow better identifcation of words}
\usage{
cleanString(stringtoclean = "")
}
\arguments{
\item{stringtoclean}{the string taken from the pdf that needs to be cleaned of problematic characters and text}
}
\description{
This function allows you to take the giant strongs ID'd with OCR and clean the text
}
\examples{
cleanString()
}
\keyword{text}
