% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameDefault.R
\name{dataFrameDefault}
\alias{dataFrameDefault}
\title{Makes an empty dataframe with col names}
\usage{
dataFrameDefault()
}
\description{
This function allows you to create a dataframe with correct column titles but empty rows.
}
\examples{
dataFrameDefault()
}
\keyword{dataframe}
