% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docTypeFinder.R
\name{docTypeFinder}
\alias{docTypeFinder}
\title{Identifies the Document Type}
\usage{
docTypeFinder(fullString = "")
}
\arguments{
\item{fullString}{the entire document}
}
\description{
This function allows you to identify which document type a pdf is
}
\examples{
docTypeFinder()
}
\keyword{identify}
