% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsiteID.R
\name{getsiteID}
\alias{getsiteID}
\title{returns the filename of each file}
\usage{
getsiteID(siteString)
}
\arguments{
\item{siteString}{the location path of the pdf}
}
\description{
This function allows you to grab the name of a site from the file
}
\examples{
getsiteID()
}
\keyword{name}
\keyword{site}
