% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr2string.R
\name{ocr2string}
\alias{ocr2string}
\title{OCR on multiple pages}
\usage{
ocr2string(pages)
}
\arguments{
\item{pages}{the images used}
}
\description{
This function allows you to modify an image and run OCR on multiple pages
}
\examples{
ocr2string()
}
\keyword{ocr}
