% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfGrabber.R
\name{pdfGrabber}
\alias{pdfGrabber}
\title{A PDF Indexer for a given folder location}
\usage{
pdfGrabber(directory)
}
\arguments{
\item{directory}{location of all your pdfs.}
}
\description{
This function allows you to graball the pdfs in a directory and returns a list.
}
\examples{
pdfGrabber()
}
\keyword{files}
