% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phraseFinder.R
\name{phraseFinder}
\alias{phraseFinder}
\title{Find the words in between two words}
\usage{
phraseFinder(stringList, firstword, lastword)
}
\arguments{
\item{stringList}{all of the page as a string list}

\item{firstword}{the first word you are searching for}

\item{lastword}{the last word you are searching for which must occur after the first word}
}
\description{
This function allows you to look for a phrase that is bounded by two words
}
\examples{
phraseFinder()
}
\keyword{find}
