% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phraseFinderAdvanced.R
\name{phraseFinderAdvanced}
\alias{phraseFinderAdvanced}
\title{Find the words in between two words}
\usage{
phraseFinderAdvanced(stringList, firstword, lastword, afterword)
}
\arguments{
\item{stringList}{all of the page as a string list}

\item{firstword}{the first word you are searching for}

\item{lastword}{the last word you are searching for which must occur after the first word}

\item{afterword}{the word that the first and last words must be after}
}
\description{
This function allows you to look for a phrase that is bounded by two words and comes after a keyword in the document
}
\examples{
phraseFinderAdvanced()
}
\keyword{finderAdvanced}
