% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phraseFinderRemover.R
\name{phraseFinderRemover}
\alias{phraseFinderRemover}
\title{Find the words in between two words}
\usage{
phraseFinderRemover(stringList, firstword, lastword, words.to.remove)
}
\arguments{
\item{stringList}{all of the page as a string list}

\item{firstword}{the first word you are searching for}

\item{lastword}{the last word you are searching for}

\item{words.to.remove}{number of words to remove from the end of a string}
}
\description{
This function allows you to look for a phrase that is bounded by two words and remove some of it
}
\examples{
phraseFinderRemover()
}
\keyword{find}
