% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType1.R
\name{processDocType1}
\alias{processDocType1}
\title{Find and store important values for document type 1}
\usage{
processDocType1(df = NULL, stringList = NULL)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 1 and save relevant information
}
\examples{
processDocType1()
}
\keyword{type1}
