% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType10.R
\name{processDocType10}
\alias{processDocType10}
\title{Find and store important values for document type 10}
\usage{
processDocType10(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 10 and save relevant information
}
\examples{
processDocType10()
}
\keyword{type10}
