% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType11.R
\name{processDocType11}
\alias{processDocType11}
\title{Find and store important values for document type 11}
\usage{
processDocType11(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 11 and save relevant information
}
\examples{
processDocType11()
}
\keyword{type10}
