% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType12.R
\name{processDocType12}
\alias{processDocType12}
\title{Find and store important values for document type 12}
\usage{
processDocType12(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 12 and save relevant information
}
\examples{
processDocType12()
}
\keyword{type10}
