% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType13.R
\name{processDocType13}
\alias{processDocType13}
\title{Find and store important values for document type 13}
\usage{
processDocType13(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 13 and save relevant information
}
\examples{
processDocType13()
}
\keyword{type10}
