% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType14.R
\name{processDocType14}
\alias{processDocType14}
\title{Find and store important values for document type 14}
\usage{
processDocType14(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 14 and save relevant information
}
\examples{
processDocType14()
}
\keyword{type10}
