% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType15.R
\name{processDocType15}
\alias{processDocType15}
\title{Find and store important values for document type 15}
\usage{
processDocType15(df = NULL, stringList = NULL)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 15 and save relevant information
}
\examples{
processDocType15()
}
\keyword{type10}
