% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType17.R
\name{processDocType17}
\alias{processDocType17}
\title{Find and store important values for document type 17}
\usage{
processDocType17(df = NULL, stringList = NULL)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 17 and save relevant information
}
\examples{
processDocType17()
}
\keyword{type10}
