% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType18.R
\name{processDocType18}
\alias{processDocType18}
\title{Find and store important values for document type 18}
\usage{
processDocType18(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 18 and save relevant information
}
\examples{
processDocType18()
}
\keyword{type18}
