% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType19.R
\name{processDocType19}
\alias{processDocType19}
\title{Find and store important values for document type 19}
\usage{
processDocType19(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 19 and save relevant information. It is very similar to type 5
}
\examples{
processDocType19()
}
\keyword{type19}
