% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType2.R
\name{processDocType2}
\alias{processDocType2}
\title{Find and store important values for document type 2}
\usage{
processDocType2(df = NULL, stringList = NULL)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 2 and save relevant information
}
\examples{
processDocType1()
}
\keyword{type1}
