% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType16.R
\name{processDocType20}
\alias{processDocType20}
\title{Find and store important values for document type 16}
\usage{
processDocType20(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 16 and save relevant information
}
\examples{
processDocType16()
}
\keyword{type16}
