% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType21.R
\name{processDocType21}
\alias{processDocType21}
\title{Find and store important values for document type 21}
\usage{
processDocType21(df = NULL, stringList = "")
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 21 and save relevant information. Doc Type 21 resembles type 5, but is for Isolates instead of Sites
}
\examples{
processDocType21()
}
\keyword{type21}
