% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType3.R
\name{processDocType3}
\alias{processDocType3}
\title{Find and store important values for document type 3}
\usage{
processDocType3(df = NULL, stringList = NULL)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 3 and save relevant information
}
\examples{
processDocType3()
}
\keyword{type3}
