% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType4.R
\name{processDocType4}
\alias{processDocType4}
\title{Find and store important values for document type 4}
\usage{
processDocType4(df = NULL, stringList = "")
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 4 and save relevant information
}
\examples{
processDocType4()
}
\keyword{type4}
