% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType5.R
\name{processDocType5}
\alias{processDocType5}
\title{Find and store important values for document type 5}
\usage{
processDocType5(df = NULL, stringList = NULL)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 5 and save relevant information
}
\examples{
processDocType5()
}
\keyword{type5}
