% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType6.R
\name{processDocType6}
\alias{processDocType6}
\title{Find and store important values for document type 6}
\usage{
processDocType6(df = NULL, stringList = NULL)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 6 and save relevant information
}
\examples{
processDocType6()
}
\keyword{type6}
