% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType7.R
\name{processDocType7}
\alias{processDocType7}
\title{Find and store important values for document type 7}
\usage{
processDocType7(df = NULL, stringList = NULL)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 7 and save relevant information
}
\examples{
processDocType7()
}
\keyword{type7}
