% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType8.R
\name{processDocType8}
\alias{processDocType8}
\title{Find and store important values for document type 8}
\usage{
processDocType8(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 8 and save relevant information
}
\examples{
processDocType8()
}
\keyword{type8}
