% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDocType9.R
\name{processDocType9}
\alias{processDocType9}
\title{Find and store important values for document type 9}
\usage{
processDocType9(df, stringList)
}
\arguments{
\item{df}{the dataframe for this site which is passed to each document type}

\item{stringList}{all of the page as a string list}
}
\description{
This function allows you to sort through document type 9 and save relevant information
}
\examples{
processDocType9()
}
\keyword{type9}
