% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPDFDirectory.R
\name{processPDFDirectory}
\alias{processPDFDirectory}
\title{Process all the PDFs in a given directory}
\usage{
processPDFDirectory(pdf_dir)
}
\arguments{
\item{pdf_dir}{the directory that has all of the PDFs you wish to process}
}
\description{
This function allows you to submit the directory of your pdfs which contain archaeological site records and returns a dataframe with the data
}
\examples{
processPDFDirectory()
}
\keyword{pdf}
