% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPDFDirectory2.R
\name{processPDFDirectory2}
\alias{processPDFDirectory2}
\title{Process all the PDFs in a given directory}
\usage{
processPDFDirectory2(pdf_dir)
}
\arguments{
\item{pdf_dir}{the directory that has all of the PDFs you wish to process}
}
\description{
This function allows you to submit the directory of your pdfs which contain archaeological site records and returns a dataframe with the data
}
\examples{
processPDFDirectory2()
}
\keyword{pdf}
