% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordFinder.R
\name{wordFinder}
\alias{wordFinder}
\title{Find the word following another word}
\usage{
wordFinder(stringList, word2find)
}
\arguments{
\item{stringList}{all of the page as a string list}

\item{word2find}{the word you are searching for}
}
\description{
This function allows you to find the word or value that comes after a word
}
\examples{
wordFinder()
}
\keyword{find}
