% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordFinderAfter.R
\name{wordFinderAfter}
\alias{wordFinderAfter}
\title{Find the word following another word}
\usage{
wordFinderAfter(stringList, word2find, number.after)
}
\arguments{
\item{stringList}{all of the page as a string list}

\item{word2find}{the word you are searching for}

\item{number.after}{the location of the word after the one identified in word2find that is being recorded}
}
\description{
This function allows you to find the word or value that comes multiple places after a word
}
\examples{
wordFinderAfter()
}
\keyword{find}
