% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanString.R
\name{cleanString}
\alias{cleanString}
\title{Text processing to allow better identifcation of words}
\usage{
cleanString(string)
}
\arguments{
\item{siteString}{the location path of the pdf}
}
\description{
This function allows you to take the giant strongs ID'd with OCR and clean the text
}
\examples{
cleanString()
}
\keyword{text}
