% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordFinderB4.R
\name{wordFinderB4}
\alias{wordFinderB4}
\title{Find the word following another word}
\usage{
wordFinderB4(stringList = NULL, word2find = NULL, numberBefore = NULL)
}
\arguments{
\item{stringList}{all of the page as a string list}

\item{word2find}{the word you are searching for}

\item{numberBefore}{the number of occurences of a string before}
}
\description{
This function allows you to find the word or value that comes before a word
}
\examples{
wordFinderB4()
}
\keyword{find}
